requirements.txt
streamlit
pdfplumber
python-docx
plotly
openai>=1.0.0
